fix price tour and hotel

+++ wp-content/themes/tagger-travel/archive-hotel.php
@@ -227,7 +227,7 @@ $exchange_rate_vnd = str_replace(',', '', $price_vnd);
                                             <?php while ($result->have_posts()) : $result->the_post();
                                                 $priceTour = get_field('hotel_price');
                                                 $price = str_replace(',','',$priceTour);
-                                                $priceAfterExchange = (int)$exchange_rate_vnd * (int)$price;
+                                                $priceAfterExchange = (float)$exchange_rate_vnd * (float)$price;
                                                 ?>
                                                 <li class="hotel_item" data-id="<?php echo get_the_ID(); ?>" data-aos="fade-up">
                                                     <div class="featured-image">
@@ -293,7 +293,7 @@ $exchange_rate_vnd = str_replace(',', '', $price_vnd);
                                             <?php while ($result->have_posts()) : $result->the_post();
                                                 $priceTour = get_field('hotel_price');
                                                 $price = str_replace(',','',$priceTour);
-                                                $priceAfterExchange = (int)$exchange_rate_vnd * (int)$price;
+                                                $priceAfterExchange = (float)$exchange_rate_vnd * (float)$price;
                                                 ?>
                                                 <li class="hotel_item" data-id="<?php echo get_the_ID(); ?>" data-aos="fade-up">
                                                     <div class="featured-image">
@@ -358,7 +358,7 @@ $exchange_rate_vnd = str_replace(',', '', $price_vnd);
                                             <?php while ($result->have_posts()) : $result->the_post();
                                                 $priceTour = get_field('hotel_price');
                                                 $price = str_replace(',','',$priceTour);
-                                                $priceAfterExchange = (int)$exchange_rate_vnd * (int)$price;
+                                                $priceAfterExchange = (float)$exchange_rate_vnd * (float)$price;
                                                 ?>
                                                 <li class="hotel_item" data-id="<?php echo get_the_ID(); ?>" data-aos="fade-up">
                                                     <div class="featured-image">
